Fix build errors and add new features (LLVM 20 / GCC 13 compat)Fix/build errors and new features

+++ include/AscendModel/IR/AscendModelInterfaces.h
@@ -14,7 +14,7 @@ namespace ascend {
 
 // Forward declarations
 class HardwareConfig;
-enum class HWUnit : uint8_t;
+enum class HWUnit : uint32_t;
 
 } // namespace ascend
 } // namespace mlir
